requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.110.3
fonttools==4.52.4
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mysql-connector-python==8.4.0
numpy==1.26.4
packaging==24.0
passlib==1.7.4
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
response==0.5.0
rsa==4.9
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
